requirements.txt
tqdm==4.50.2
Flask==1.1.2
requests==2.24.0
Pillow==8.2.0
